Implement YAML block additional feature
This implements support for parsing YAML blocks and front-matter
following the Pandoc rules:

  1. Delimited by three hyphens (---) at the top
  2. Delimited by three hyphens (---) or dots (...) at the bottom
  3. May occur anywhere in the document, but if not at the beginning
     must be preceded by a blank line.

BlockTag.YamlBlock is defined, and FencedCodeData is used to store the
closing fence character (either `-` or `.`) and to indicate whether the
closing fence has been seen in the same manner as BlockTag.FencedCode.

YAML support may be enabled via two-flavors:

  1. CommonMarkAdditionalFeatures.YamlBlocks: this allows for any
     number of YAML blocks anywhere in the document.

  2. CommonMarkAdditionalFeatures.YamlFrontMatterOnly: allows for
     exactly one YAML block, defined on the first line of the document.

The HTML formatters treat BlockTag.YamlBlock blocks the same way as
BlockTag.FencedCode blocks, except that instead of writing the info
string, a `class="language-yaml"` attribute will be written.